refactor: drivers in dedicated packages
