+++ src/dip_coater/motor_driver/tmc2660/__init__.py
@@ -1,0 +1,7 @@
+from dip_coater.motor_driver.tmc2660.tmc2660 import (
+    ChopperMode,
+    MotorDriverTMC2660,
+    StepDirSource,
+    VSenseFullScale,
+)
+from dip_coater.logging.tmc2660_logger import TMC2660LogLevel
