+++ src/dip_coater/motor_driver/tmc2660/tmc2660.py
@@ -19,7 +19,7 @@ except ModuleNotFoundError:
 
 from dip_coater.logging.tmc2660_logger import TMC2660Logger, TMC2660LogLevel
 from dip_coater.motor_driver.motor_driver_interface import MotorDriver
-from dip_coater.motor_driver.tmc2660_dummy import (
+from dip_coater.motor_driver.tmc2660.tmc2660_dummy import (
     DummyEvalBoard,
     DummyInterface,
     DummyLandungsbruecke,
